separated concerns
