--- Test.md
@@ -1,36 +1,0 @@
-# My First Blog Post
-
-This is a **basic markdown file** with *various formatting* options.
-
-## Introduction
-
-Welcome to my blog! This post demonstrates ~~some~~ **most** of the basic Markdown syntax.
-
-### What You'll Learn
-
-- How to format text
-- Basic *emphasis* and **strong** text
-- `Inline code` examples
-- TESTING TESTING TESTING!!!
-
-#### Code Example
-
-Here's some `inline code` and a code block:
-
-```go
-func main() {
-    fmt.Println("Hello, Markdown!")
-}
-```
-
-##### Quote Section
-
-> This is a blockquote.
-> It can span multiple lines.
-
-###### Final Notes
-
-Regular paragraph text here. You can use **bold**, *italic*, and even `code` within sentences.
-
-Another paragraph with a [link to Google](https://google.com).
-
