--- Test2.md
@@ -1,82 +1,0 @@
-# Basic Lists
-
-## Unordered List
-- First item
-- Second item
-- Third item
-
-## Ordered List
-1. Step one
-2. Step two
-3. Step three
-
-# Mixed Content
-
-This is a paragraph before the list.
-
-- Item with **bold** text
-- Item with *italic* text
-- Item with `code` text
-
-Another paragraph here.
-
-1. Numbered item with **formatting**
-2. Another numbered item
-3. Final numbered item
-
-End paragraph.
-
-# Multiple Lists
-
-First list:
-- Apple
-- Banana
-- Cherry
-
-Some text between lists.
-
-Second list:
-1. Red
-2. Green
-3. Blue
-
-Another unordered list:
-- Dog
-- Cat
-- Bird
-
-# Lists with Spacing
-
-- First item
-
-- Second item after empty line
-
-- Third item
-
-1. Numbered first
-
-2. Numbered second
-
-3. Numbered third
-
-
-# Complex Lists
-
-## Shopping List
-- Groceries
-- Hardware items
-- Books
-
-## To-Do Today
-1. Wake up
-2. Drink coffee
-3. Code SSG
-4. Test lists
-5. Debug issues
-
-## Random Items
-- Item with ~~strikethrough~~
-- Item with **bold** and *italic*
-- Item with `inline code`
-
-That's all!
