--- index.html
@@ -1,125 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-</head>
-<body>
-    <h1>Basic Lists</h1>
-
-<h2>Unordered List</h2>
-<ul>
-  <li>First item</li>
-  <li>Second item</li>
-  <li>Third item</li>
-</ul>
-
-<h2>Ordered List</h2>
-<ol>
-  <li>Step one</li>
-  <li>Step two</li>
-  <li>Step three</li>
-</ol>
-
-<h1>Mixed Content</h1>
-
-<p>This is a paragraph before the list.</p>
-
-<ul>
-  <li>Item with <strong>bold</strong> text</li>
-  <li>Item with <em>italic</em> text</li>
-  <li>Item with <code>code</code> text</li>
-</ul>
-
-<p>Another paragraph here.</p>
-
-<ol>
-  <li>Numbered item with <strong>formatting</strong></li>
-  <li>Another numbered item</li>
-  <li>Final numbered item</li>
-</ol>
-
-<p>End paragraph.</p>
-
-<h1>Multiple Lists</h1>
-
-<p>First list:</p>
-<ul>
-  <li>Apple</li>
-  <li>Banana</li>
-  <li>Cherry</li>
-</ul>
-
-<p>Some text between lists.</p>
-
-<p>Second list:</p>
-<ol>
-  <li>Red</li>
-  <li>Green</li>
-  <li>Blue</li>
-</ol>
-
-<p>Another unordered list:</p>
-<ul>
-  <li>Dog</li>
-  <li>Cat</li>
-  <li>Bird</li>
-</ul>
-
-<h1>Lists with Spacing</h1>
-
-<ul>
-  <li>First item</li>
-</ul>
-
-<ul>
-  <li>Second item after empty line</li>
-</ul>
-
-<ul>
-  <li>Third item</li>
-</ul>
-
-<ol>
-  <li>Numbered first</li>
-</ol>
-
-<ol>
-  <li>Numbered second</li>
-</ol>
-
-<ol>
-  <li>Numbered third</li>
-</ol>
-
-
-<h1>Complex Lists</h1>
-
-<h2>Shopping List</h2>
-<ul>
-  <li>Groceries</li>
-  <li>Hardware items</li>
-  <li>Books</li>
-</ul>
-
-<h2>To-Do Today</h2>
-<ol>
-  <li>Wake up</li>
-  <li>Drink coffee</li>
-  <li>Code SSG</li>
-  <li>Test lists</li>
-  <li>Debug issues</li>
-</ol>
-
-<h2>Random Items</h2>
-<ul>
-  <li>Item with <del>strikethrough</del></li>
-  <li>Item with <strong>bold</strong> and <em>italic</em></li>
-  <li>Item with <code>inline code</code></li>
-</ul>
-
-<p>That's all!</p>
-
-</body>
-</html>
