+++ internal/types/models.go
@@ -1,0 +1,12 @@
+package types
+
+
+type TEXT struct {
+	Form        string
+	Sliced      []string
+	HasChildren bool
+}
+
+type HEADER struct {
+	Form string
+}
