+++ public/Test.md
@@ -1,0 +1,133 @@
+# The Ultimate ***Complex*** Test Document
+
+This **document** contains ***extremely complex*** *formatting* **with** ~~multiple~~ **nested** *formatting* `inline code` elements.
+
+## Deeply Nested Lists
+
+- First level with **bold text**
+- Item with *italic* and `code` combined
+- Complex item with ~~strikethrough~~ and ***bold italic***
+- Code sample: `function(param1, param2)` with explanation
+
+1. Numbered item with **formatting**
+2. Item containing `inline code` and *emphasis*
+3. Multiple ~~deletions~~ and **additions** in one line
+4. Complex: ***bold italic*** with `code` and ~~strike~~
+
+### Mixed Content Sections
+
+Paragraph with **bold**, *italic*, ~~strikethrough~~, and `code` all mixed together.
+
+- List item with **bold *nested italic* text**
+- Item with `code containing **bold** inside`
+- Complex: ~~strike with *italic* inside~~
+
+#### Code Blocks with Different Languages
+
+Here's a Go code example:
+
+```go
+func complexFunction(param string) error {
+    if param == "" {
+        return errors.New("empty parameter")
+    }
+    return nil
+}
+```
+
+And here's some JavaScript:
+
+```javascript
+const complexObject = {
+    method: function(arg) {
+        return arg.toUpperCase();
+    },
+    nested: {
+        value: 42
+    }
+};
+```
+
+##### Blockquotes and Mixed Formatting
+
+> This is a blockquote with **bold text**
+> And *italic formatting* mixed in
+> Even `inline code` works here
+
+###### Ultra Complex Combinations
+
+This paragraph has ***bold italic***, ~~strikethrough with **bold inside**~~, `code with *italic*`, and regular text all mixed.
+
+- Item 1: **Bold with `code` inside**
+- Item 2: *Italic with ~~strike~~ nested*
+- Item 3: `Code with **bold** and *italic*`
+- Item 4: ~~Strike with ***bold italic*** inside~~
+
+1. Complex formatting: ***`code in bold italic`***
+2. Multiple layers: **bold with *italic and `code`* inside**
+3. Strike combo: ~~deleted **bold** and *italic*~~
+
+## Extreme Edge Cases
+
+Text with **bold containing *italic with `code` inside* text** complexity.
+
+- **Bold *italic ~~strike `code`~~ text* end**
+- Multiple `code` segments in **bold *italic* text**
+- Nested: ~~strike **bold *italic `code` text* bold** strike~~
+
+### Code Without Language
+
+```
+// This is a plain code block
+function generic() {
+    console.log("No specific language");
+}
+```
+
+#### Final Complex Section
+
+The ultimate test: ***Bold italic*** with ~~strikethrough containing **bold** and *italic*~~ plus `inline code with **formatting**` all in one paragraph.
+
+1. ***`Bold italic code`*** combination
+2. **Bold with ~~strike *italic*~~ nested**
+3. `Code containing ***all*** formatting`
+4. ~~Strike with **bold `code` bold** strike~~
+
+End of complex document with **final** *formatting* ~~test~~ and `code`.
+
+---
+
+## Horizontal Rule Test
+
+Text before horizontal rule.
+
+---
+
+Text after horizontal rule.
+
+## List Spacing Tests
+
+- Item 1
+
+- Item 2 after empty line
+
+- Item 3
+
+Some text between lists.
+
+1. Numbered 1
+
+2. Numbered 2
+
+3. Numbered 3
+
+## Multiple Empty Lines Test
+
+
+
+Multiple empty lines above this text.
+
+
+Single empty line above.
+
+No empty line above.
