+++ src/index.html
@@ -1,0 +1,176 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Document</title>
+</head>
+<body>
+    <h1>The Ultimate <em><strong>Complex</strong></em> Test Document</h1>
+
+<p>This <strong>document</strong> contains <em><strong>extremely complex</strong></em> <em>formatting</em> <strong>with</strong> <del>multiple</del> <strong>nested</strong> <em>formatting</em> <code>inline code</code> elements.</p>
+
+<h2>Deeply Nested Lists</h2>
+
+<ul>
+  <li>First level with <strong>bold text</strong></li>
+  <li>Item with <em>italic</em> and <code>code</code> combined</li>
+  <li>Complex item with <del>strikethrough</del> and <em><strong>bold italic</strong></em></li>
+  <li>Code sample: <code>function(param1, param2)</code> with explanation</li>
+</ul>
+
+<ol>
+  <li>Numbered item with <strong>formatting</strong></li>
+  <li>Item containing <code>inline code</code> and <em>emphasis</em></li>
+  <li>Multiple <del>deletions</del> and <strong>additions</strong> in one line</li>
+  <li>Complex: <em><strong>bold italic</strong></em> with <code>code</code> and <del>strike</del></li>
+</ol>
+
+<h3>Mixed Content Sections</h3>
+
+<p>Paragraph with <strong>bold</strong>, <em>italic</em>, <del>strikethrough</del>, and <code>code</code> all mixed together.</p>
+
+<ul>
+  <li>List item with <strong>bold <em>nested italic</em> text</strong></li>
+  <li>Item with <code>code containing <strong>bold</strong> inside</code></li>
+  <li>Complex: <del>strike with <em>italic</em> inside</del></li>
+</ul>
+
+<h4>Code Blocks with Different Languages</h4>
+
+<p>Here's a Go code example:</p>
+
+<pre>
+ <code class="go">
+ func complexFunction(param string) error {
+    if param == "" {
+        return errors.New("empty parameter")
+    }
+    return nil
+}
+ </code>
+</pre>
+
+<p>And here's some JavaScript:</p>
+
+<pre>
+ <code class="javascript">
+ const complexObject = {
+    method: function(arg) {
+        return arg.toUpperCase();
+    },
+    nested: {
+        value: 42
+    }
+};
+ </code>
+</pre>
+
+<h5>Blockquotes and Mixed Formatting</h5>
+
+<p>> This is a blockquote with <strong>bold text</strong></p>
+<p>> And <em>italic formatting</em> mixed in</p>
+<p>> Even <code>inline code</code> works here</p>
+
+<h6>Ultra Complex Combinations</h6>
+
+<p>This paragraph has <em><strong>bold italic</strong></em>, <del>strikethrough with <strong>bold inside</strong></del>, <code>code with <em>italic</em></code>, and regular text all mixed.</p>
+
+<ul>
+  <li>Item 1: <strong>Bold with <code>code</code> inside</strong></li>
+  <li>Item 2: <em>Italic with <del>strike</del> nested</em></li>
+  <li>Item 3: <code>Code with <strong>bold</strong> and <em>italic</em></code></li>
+  <li>Item 4: <del>Strike with <em><strong>bold italic</strong></em> inside</del></li>
+</ul>
+
+<ol>
+  <li>Complex formatting: <em><strong><code>code in bold italic</code></strong></em></li>
+  <li>Multiple layers: <strong>bold with <em>italic and <code>code</code></em> inside</strong></li>
+  <li>Strike combo: <del>deleted <strong>bold</strong> and <em>italic</em></del></li>
+</ol>
+
+<h2>Extreme Edge Cases</h2>
+
+<p>Text with <strong>bold containing <em>italic with <code>code</code> inside</em> text</strong> complexity.</p>
+
+<ul>
+  <li><strong>Bold <em>italic <del>strike <code>code</code></del> text</em> end</strong></li>
+  <li>Multiple <code>code</code> segments in <strong>bold <em>italic</em> text</strong></li>
+  <li>Nested: <del>strike <strong>bold <em>italic <code>code</code> text</em> bold</strong> strike</del></li>
+</ul>
+
+<h3>Code Without Language</h3>
+
+<pre>
+ <code class="">
+ // This is a plain code block
+function generic() {
+    console.log("No specific language");
+}
+ </code>
+</pre>
+
+<h4>Final Complex Section</h4>
+
+<p>The ultimate test: <em><strong>Bold italic</strong></em> with <del>strikethrough containing <strong>bold</strong> and <em>italic</em></del> plus <code>inline code with <strong>formatting</strong></code> all in one paragraph.</p>
+
+<ol>
+  <li><em><strong><code>Bold italic code</code></strong></em> combination</li>
+  <li><strong>Bold with <del>strike <em>italic</em></del> nested</strong></li>
+  <li><code>Code containing <em><strong>all</strong></em> formatting</code></li>
+  <li><del>Strike with <strong>bold <code>code</code> bold</strong> strike</del></li>
+</ol>
+
+<p>End of complex document with <strong>final</strong> <em>formatting</em> <del>test</del> and <code>code</code>.</p>
+
+<p>---</p>
+
+<h2>Horizontal Rule Test</h2>
+
+<p>Text before horizontal rule.</p>
+
+<p>---</p>
+
+<p>Text after horizontal rule.</p>
+
+<h2>List Spacing Tests</h2>
+
+<ul>
+  <li>Item 1</li>
+</ul>
+
+<ul>
+  <li>Item 2 after empty line</li>
+</ul>
+
+<ul>
+  <li>Item 3</li>
+</ul>
+
+<p>Some text between lists.</p>
+
+<ol>
+  <li>Numbered 1</li>
+</ol>
+
+<ol>
+  <li>Numbered 2</li>
+</ol>
+
+<ol>
+  <li>Numbered 3</li>
+</ol>
+
+<h2>Multiple Empty Lines Test</h2>
+
+
+
+<p>Multiple empty lines above this text.</p>
+
+
+<p>Single empty line above.</p>
+
+<p>No empty line above.</p>
+
+</body>
+</html>
